requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
django-redis
Pillow
pytest
pytest-django
python-dotenv
django-cors-headers
django-filter
daphne            # The Interface Server
channels          # The WebSocket Framework
channels_redis    # The Communication Layer
tensorflow-cpu
numpy
django-debug-toolbar
django-prometheus
strawberry-graphql-django
gunicorn           # Production WSGI/ASGI Server
uvicorn            # ASGI Worker for Gunicorn
flake8             # Linter (used in CI pipeline)